pyproject.toml
[tool.poetry]
name = "kaggle_academics"
version = "0.1.0"
description = ""
authors = ["ednguyenn"]

[tool.poetry.dependencies]
python = "3.11.5"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
pandas = "^2.2.2"
matplotlib = "^3.9.0"
scikit-learn = "^1.5.0"
seaborn = "^0.13.2"
plotly = "^5.22.0"
nbformat = "^4.2.0"
dill = "^0.3.8"
xgboost = "^2.0.3"
hyperopt = "^0.2.7"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.mccabe]
max-complexity = 10
